+++ Services/Services.Contracts/GroupMember/GroupMemberFilterDto.cs
@@ -2,8 +2,8 @@
 
 public class GroupMemberFilterDto
 {
-    public int GroupId { get; set; }
-    public int UserId { get; set; }
+    public long GroupId { get; set; }
+    public long UserId { get; set; }
 
     public int ItemsPerPage { get; set; }
