+++ Services/Services.Contracts/GroupMember/UpdatingGroupMemberDto.cs
@@ -2,7 +2,7 @@
 
 public class UpdatingGroupMemberDto
 {
-    public int GroupId { get; set; }
-    public int UserId { get; set; }
+    public long GroupId { get; set; }
+    public long UserId { get; set; }
     public string Role { get; set; }
 }
