+++ Services/Services.Contracts/SearchGroup/CreatingSearchGroupDto.cs
@@ -2,6 +2,6 @@
 
 public class CreatingSearchGroupDto
 {
-    public int RequestId { get; set; }
-    public int LeaderId { get; set; }
+    public long RequestId { get; set; }
+    public long LeaderId { get; set; }
 }
